conversions module rejiggered -- needs work
